Separated opentracing mock tracer tests from opentracing api compatibility tests.
